+++ packages/kit/src/extensions/white-ink/index.ts
@@ -1,4 +1,6 @@
+export * from "./capability";
 export * from "./WhiteInkTool";
+export * from "./WhiteInkCapabilityExtension";
 export * from "./commands";
 export * from "./config";
 export * from "./model";